feat: use unmasked int type in function calls

+++ b-tests/LCSF_Bridge_Test.h
@@ -181,13 +181,13 @@ enum _lcsf_test_cc6_att_id {
 // --- Public Function Prototypes ---
 
 /**
- * \fn bool LCSF_Bridge_TestInit(uint16_t filoSize)
+ * \fn bool LCSF_Bridge_TestInit(size_t filoSize)
  * \brief Initialize the module
  *
  * \param filoSize size of the module filo (number of element)
  * \return bool: true if operation was a success
  */
-bool LCSF_Bridge_TestInit(uint16_t filoSize);
+bool LCSF_Bridge_TestInit(size_t filoSize);
 
 /**
  * \fn bool LCSF_Bridge_TestReceive(lcsf_valid_cmd_t *pValidCmd)
@@ -199,14 +199,14 @@ bool LCSF_Bridge_TestInit(uint16_t filoSize);
 bool LCSF_Bridge_TestReceive(lcsf_valid_cmd_t *pValidCmd);
 
 /**
- * \fn bool LCSF_Bridge_TestSend(uint16_t cmdName, test_cmd_payload_t *pCmdPayload)
+ * \fn bool LCSF_Bridge_TestSend(uint_fast16_t cmdName, test_cmd_payload_t *pCmdPayload)
  * \brief Receive command from Test_Main and transmit to LCSF_Validator
  *
  * \param cmdName name of the command
  * \param pValidCmd pointer to the valid command
  * \return bool: true if operation was a success
  */
-bool LCSF_Bridge_TestSend(uint16_t cmdName, test_cmd_payload_t *pCmdPayload);
+bool LCSF_Bridge_TestSend(uint_fast16_t cmdName, test_cmd_payload_t *pCmdPayload);
 
 // *** End Definitions ***
 #endif // Lcsf_bridge_test_h
