+++ src/protocol/LCSF_Bridge_Example.h
@@ -76,13 +76,13 @@ enum _lcsf_example_att_id {
 
 // --- Public Function Prototypes ---
 /**
- * \fn bool LCSF_Bridge_ExampleInit(uint16_t filoSize)
+ * \fn bool LCSF_Bridge_ExampleInit(size_t filoSize)
  * \brief Initialize the module
  *
  * \param filoSize size of the module filo (number of element)
  * \return bool: true if operation was a success
  */
-bool LCSF_Bridge_ExampleInit(uint16_t filoSize);
+bool LCSF_Bridge_ExampleInit(size_t filoSize);
 
 /**
  * \fn bool LCSF_Bridge_ExampleReceive(lcsf_valid_cmd_t *pValidCmd)
@@ -94,14 +94,14 @@ bool LCSF_Bridge_ExampleInit(uint16_t filoSize);
 bool LCSF_Bridge_ExampleReceive(lcsf_valid_cmd_t *pValidCmd);
 
 /**
- * \fn bool LCSF_Bridge_ExampleSend(uint16_t cmdName, void *pCmdPayload)
+ * \fn bool LCSF_Bridge_ExampleSend(uint_fast16_t cmdName, void *pCmdPayload)
  * \brief Receive command from Example_Main and transmit to LCSF_Validator
  *
  * \param cmdName name of the command
  * \param pValidCmd pointer to the valid command
  * \return bool: true if operation was a success
  */
-bool LCSF_Bridge_ExampleSend(uint16_t cmdName, example_cmd_payload_t *pCmdPayload);
+bool LCSF_Bridge_ExampleSend(uint_fast16_t cmdName, example_cmd_payload_t *pCmdPayload);
 
 // *** End Definitions ***
 #endif // LCSF_Bridge_Example_h
